eval/requirements.txt
# Agent dependencies (shared)
atproto>=0.0.54
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
openai>=1.50.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Eval-specific dependencies
sentence-transformers>=3.0.0
torch>=2.0.0
numpy>=1.26.0
